requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
statsmodels
requests
python-dotenv
pandas-datareader
jupyter
notebook
ipykernel
openpyxl
python-docx